Airflow/requirements.txt
apache-airflow-providers-apache-hdfs==4.11.1
apache-airflow-providers-apache-spark==5.4.1
apache-airflow-providers-common-compat==1.10.1
apache-airflow-providers-http==4.10.0
requests==2.31.0